requirements.txt
Django==3.2.5
django-filter==21.1
django-rest-knox==4.1.0
djangorestframework==3.12.4
markdown==3.3.4